move font config
